+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add implementation function for `dict.pop` (#517)
+- Remove `WeakExtension` (#517)
 - Fix propagation of no-return-unless constraints from calls
   to unions (#518)
 - Initial support for variable-length heterogeneous sequences